# [FLINK-14346] [serialization] faster implementation of StringValue writeString and readString