--- a/flink-core/src/test/java/org/apache/flink/types/StringSerializationTest.java
+++ b/flink-core/src/test/java/org/apache/flink/types/StringSerializationTest.java
@@ -20,18 +20,22 @@ package org.apache.flink.types;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
+import static org.junit.Assert.assertArrayEquals;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
-import java.util.Random;
+import java.io.DataInput;
+import java.io.DataOutput;
 
-import org.apache.flink.types.StringValue;
+import java.util.Random;
+import java.util.function.BiConsumer;
+import java.util.function.Function;
+
 import org.apache.flink.util.StringUtils;
 import org.junit.Test;
-
 
 /**
  * Test for the serialization of Strings through the StringValue class.
@@ -70,6 +74,41 @@ public class StringSerializationTest {
 			};
 
 			testSerialization(testStrings);
+		}
+		catch (Exception e) {
+			System.err.println(e.getMessage());
+			e.printStackTrace();
+			fail("Exception in test: " + e.getMessage());
+		}
+	}
+
+	@Test
+	public void testStringBinaryCompatibility() {
+		try {
+			String[] testStrings = new String[] {
+				StringUtils.getRandomString(rnd, 10000, 1024 * 1024 * 2, (char) 1, (char) 127),
+				StringUtils.getRandomString(rnd, 10000, 1024 * 1024 * 2, (char) 128, (char) 16383),
+				StringUtils.getRandomString(rnd, 10000, 1024 * 1024 * 2, (char) 16384, (char) 65535),
+				StringUtils.getRandomString(rnd, 10000, 1024 * 1024 * 2, (char) 1, (char) 16383),
+				StringUtils.getRandomString(rnd, 10000, 1024 * 1024 * 2, (char) 1, (char) 65535),
+				StringUtils.getRandomString(rnd, 10000, 1024 * 1024 * 2, (char) 128, (char) 65535)
+			};
+
+			for (String testString: testStrings) {
+				// new and old impl should produce the same binary result
+				byte[] oldBytes = serializeBytes(testString, StringSerializationTest::oldWriteString);
+				byte[] newBytes = serializeBytes(testString, StringSerializationTest::newWriteString);
+				assertArrayEquals(oldBytes, newBytes);
+				// old impl should read bytes from new one
+				String oldString = deserializeBytes(newBytes, StringSerializationTest::oldReadString);
+				assertEquals(oldString, testString);
+				// new impl should read bytes from old one
+				String newString = deserializeBytes(oldBytes, StringSerializationTest::newReadString);
+				assertEquals(newString, testString);
+				// it should roundtrip over new impl
+				String roundtrip = deserializeBytes(newBytes, StringSerializationTest::newReadString);
+				assertEquals(roundtrip, testString);
+			}
 		}
 		catch (Exception e) {
 			System.err.println(e.getMessage());
@@ -162,6 +201,22 @@ public class StringSerializationTest {
 			fail("Exception in test: " + e.getMessage());
 		}
 	}
+
+	public static final byte[] serializeBytes(String value, BiConsumer<String, DataOutput> writer) throws IOException {
+		ByteArrayOutputStream buffer = new ByteArrayOutputStream();
+		DataOutputStream stream = new DataOutputStream(buffer);
+		writer.accept(value, stream);
+		stream.close();
+		return buffer.toByteArray();
+	}
+
+	public static final String deserializeBytes(byte[] value, Function<DataInput, String> reader) throws IOException {
+		ByteArrayInputStream buffer = new ByteArrayInputStream(value);
+		DataInputStream stream = new DataInputStream(buffer);
+		String result = reader.apply(stream);
+		stream.close();
+		return result;
+	}
 	
 	public static final void testSerialization(String[] values) throws IOException {
 		ByteArrayOutputStream baos = new ByteArrayOutputStream(4096);
@@ -221,5 +276,107 @@ public class StringSerializationTest {
 		
 		assertEquals("Wrong number of deserialized values", values.length, num);
 	}
-	
+
+	// needed to test the binary compatibility for new/old string serialization code
+	private static final int HIGH_BIT = 0x1 << 7;
+	private static String oldReadString(DataInput in) {
+		try {
+			// the length we read is offset by one, because a length of zero indicates a null value
+			int len = in.readUnsignedByte();
+
+			if (len == 0) {
+				return null;
+			}
+
+			if (len >= HIGH_BIT) {
+				int shift = 7;
+				int curr;
+				len = len & 0x7f;
+				while ((curr = in.readUnsignedByte()) >= HIGH_BIT) {
+					len |= (curr & 0x7f) << shift;
+					shift += 7;
+				}
+				len |= curr << shift;
+			}
+
+			// subtract one for the null length
+			len -= 1;
+
+			final char[] data = new char[len];
+
+			for (int i = 0; i < len; i++) {
+				int c = in.readUnsignedByte();
+				if (c < HIGH_BIT) {
+					data[i] = (char) c;
+				} else {
+					int shift = 7;
+					int curr;
+					c = c & 0x7f;
+					while ((curr = in.readUnsignedByte()) >= HIGH_BIT) {
+						c |= (curr & 0x7f) << shift;
+						shift += 7;
+					}
+					c |= curr << shift;
+					data[i] = (char) c;
+				}
+			}
+
+			return new String(data, 0, len);
+		} catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	private static final void oldWriteString(CharSequence cs, DataOutput out) {
+		try {
+			if (cs != null) {
+				// the length we write is offset by one, because a length of zero indicates a null value
+				int lenToWrite = cs.length()+1;
+				if (lenToWrite < 0) {
+					throw new IllegalArgumentException("CharSequence is too long.");
+				}
+
+				// write the length, variable-length encoded
+				while (lenToWrite >= HIGH_BIT) {
+					out.write(lenToWrite | HIGH_BIT);
+					lenToWrite >>>= 7;
+				}
+				out.write(lenToWrite);
+
+				// write the char data, variable length encoded
+				for (int i = 0; i < cs.length(); i++) {
+					int c = cs.charAt(i);
+
+					while (c >= HIGH_BIT) {
+						out.write(c | HIGH_BIT);
+						c >>>= 7;
+					}
+					out.write(c);
+				}
+			} else {
+				out.write(0);
+			}
+		} catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	public static String newReadString(DataInput in) {
+		try {
+			return StringValue.readString(in);
+		} catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	public static void newWriteString(CharSequence cs, DataOutput out) {
+		try {
+			StringValue.writeString(cs, out);
+		} catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+
+
 }