--- a/flink-core/src/test/java/org/apache/flink/types/StringSerializationTest.java
+++ b/flink-core/src/test/java/org/apache/flink/types/StringSerializationTest.java
@@ -30,7 +30,6 @@ import java.io.IOException;
 import java.io.DataInput;
 import java.io.DataOutput;
 
-import java.util.ArrayList;
 import java.util.Random;
 import java.util.function.BiConsumer;
 import java.util.function.Function;