@shuttie 2019-12-03
Yep, there is a regression for strings shorter than 6 characters, see this more granular benchmark:

```

Before fallback:

[info] Benchmark                                    (length)  (stringType)  Mode  Cnt   Score    Error  Units
[info] StringSerializerBenchmark.serializeDefault          1         ascii  avgt    5  33.383 ±  2.796  ns/op
[info] StringSerializerBenchmark.serializeDefault          2         ascii  avgt    5  32.731 ±  2.470  ns/op
[info] StringSerializerBenchmark.serializeDefault          3         ascii  avgt    5  37.619 ±  3.950  ns/op
[info] StringSerializerBenchmark.serializeDefault          4         ascii  avgt    5  42.452 ±  3.703  ns/op
[info] StringSerializerBenchmark.serializeDefault          5         ascii  avgt    5  46.887 ±  2.906  ns/op
[info] StringSerializerBenchmark.serializeDefault          6         ascii  avgt    5  57.461 ± 14.265  ns/op
[info] StringSerializerBenchmark.serializeDefault          7         ascii  avgt    5  58.337 ±  2.813  ns/op
[info] StringSerializerBenchmark.serializeImproved         1         ascii  avgt    5  37.015 ± 11.327  ns/op
[info] StringSerializerBenchmark.serializeImproved         2         ascii  avgt    5  40.723 ±  9.182  ns/op
[info] StringSerializerBenchmark.serializeImproved         3         ascii  avgt    5  43.556 ± 10.250  ns/op
[info] StringSerializerBenchmark.serializeImproved         4         ascii  avgt    5  48.410 ± 12.323  ns/op
[info] StringSerializerBenchmark.serializeImproved         5         ascii  avgt    5  47.770 ±  7.285  ns/op
[info] StringSerializerBenchmark.serializeImproved         6         ascii  avgt    5  48.477 ±  7.607  ns/op
[info] StringSerializerBenchmark.serializeImproved         7         ascii  avgt    5  49.082 ± 13.026  ns/op

After fallback:

[info] Benchmark                                    (length)  (stringType)  Mode  Cnt   Score   Error  Units
[info] StringSerializerBenchmark.serializeImproved         1         ascii  avgt    5  31.794 ± 0.898  ns/op
[info] StringSerializerBenchmark.serializeImproved         2         ascii  avgt    5  30.904 ± 0.814  ns/op
[info] StringSerializerBenchmark.serializeImproved         3         ascii  avgt    5  35.260 ± 1.481  ns/op
[info] StringSerializerBenchmark.serializeImproved         4         ascii  avgt    5  40.210 ± 1.505  ns/op
[info] StringSerializerBenchmark.serializeImproved         5         ascii  avgt    5  45.301 ± 2.434  ns/op
[info] StringSerializerBenchmark.serializeImproved         6         ascii  avgt    5  43.255 ± 8.550  ns/op
[info] StringSerializerBenchmark.serializeImproved         7         ascii  avgt    5  45.846 ± 7.652  ns/op
```